+++ packages/web-ui/CHANGELOG.md
@@ -12,6 +12,18 @@
 
 ### Removed
 
+## [2026.7.9] - 2026-07-09
+
+### Breaking Changes
+
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [2026.7.5-2] - 2026-07-05
 
 ### Breaking Changes